refactor: splitt viewModels.ts i domenespesifikke moduler (#167)
2199-linjersfilen er delt inn i 7 fokuserte moduler under view-models/:
_helpers.ts (488), skills.ts (946), professions.ts (333), societies.ts (386),
access.ts (51), overview.ts (69), audit.ts (1).
viewModels.ts er nå en kompatibel re-eksport-barrel — ingen importsteder er endret.
