+++ apps/web/src/lib/admin/view-models/audit.ts
@@ -1,0 +1,1 @@
+export type AuditSeverity = "blocking" | "info" | "warning";
