+++ apps/web/src/lib/admin/view-models/index.ts
@@ -1,0 +1,6 @@
+export type { AuditSeverity } from "./audit";
+export * from "./skills";
+export * from "./professions";
+export * from "./societies";
+export * from "./access";
+export * from "./overview";
